+++ django_cachex/serializers/__init__.py
@@ -1,4 +1,3 @@
-# Serializers module
 from django_cachex.serializers.base import BaseSerializer
 
 __all__ = ["BaseSerializer"]